re-add test datasets that were purged during large file cleanup
The rss php uses the local dataset's filesystem timestamp to
set the 'pubdate' field.  Git does not preserve file modification times.
Therefore, git will screw up pubdates if we try to track those
files in the repo.

If we just using 'git rm' on the files it would leave the large blobs
in the history, slowing down clone operations, etc.

To avoid that, I purged the datasets from the history using
git-filter-repo.  Of the various tools available, git-filter-repo
provides the best path glob support.

Because the test datasets were removed as well, I am
re-adding them to the repo with this commit.